+++ Changelog.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Added
+- Sentry integration
+
+### Changed
+- Simplify logging
+
 ## [0.2.0] - 2022-03-03
 ### Added
 - Changelog
@@ -22,6 +29,6 @@ All notable changes to this project will be documented in this file.
 ## [0.1.0] - 2021-03-30
 -  Presentation and content APIs
 
-
+[Unreleased]: https://bbpgitlab.epfl.ch/viz/veeone/veedrive/-/compare/0.2.0...master?from_project_id=124
 [0.2.0]: https://bbpgitlab.epfl.ch/viz/veeone/veedrive/-/compare/0.1.0...0.2.0?from_project_id=124
 [0.1.0]: https://bbpgitlab.epfl.ch/viz/veeone/veedrive/-/compare/e7c9bc69247b67440166d754738c0ccc42d55cbb...0.1.0?from_project_id=124